Remove SharedProject and add a breaking change notice
