+++ vnext/Microsoft.ReactNative.Managed.CodeGen/Microsoft.ReactNative.Managed.CodeGen.csproj
@@ -37,7 +37,7 @@
       But we don't want to actually link against the Microsfot.ReactNative.Managed assembly becuase then we would have a cyclic dependency
       to compile-time generate the state for that assembly.
     -->
-    <Compile Include="..\Microsoft.ReactNative.SharedManaged\ReactAttributes.cs" Link="ReactAttributes.cs" />
+    <Compile Include="..\Microsoft.ReactNative.Managed\ReactAttributes.cs" Link="ReactAttributes.cs" />
   </ItemGroup>
 
   <ItemGroup>
