+++ vnext/Microsoft.ReactNative.Managed/AttributedViewManager.cs
