+++ vnext/Microsoft.ReactNative.Managed/IJSValueTreeReader.cs
