+++ vnext/Microsoft.ReactNative.Managed/IReactPromise.cs
