+++ vnext/Microsoft.ReactNative.Managed/JSValue.cs
