+++ vnext/Microsoft.ReactNative.Managed/JSValueReaderGenerator.cs
