+++ vnext/Microsoft.ReactNative.Managed/JSValueTreeReader.cs
