+++ vnext/Microsoft.ReactNative.Managed/JSValueTreeWriter.cs
