+++ vnext/Microsoft.ReactNative.Managed/JSValueWriter.cs
