+++ vnext/Microsoft.ReactNative.Managed/JSValueWriterGenerator.cs
