+++ vnext/Microsoft.ReactNative.Managed/Microsoft.ReactNative.Managed.csproj
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="16.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="16.0" DefaultTargets="Build" 
+  xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -101,8 +102,40 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="AttributedViewManager.cs" /> 
+    <Compile Include="IJSValueTreeReader.cs" />
+    <Compile Include="IReactPromise.cs" />
+    <Compile Include="JSValue.cs" />
+    <Compile Include="JSValueGenerator.cs" />
+    <Compile Include="JSValueReader.cs" />
+    <Compile Include="JSValueReaderGenerator.cs" />
+    <Compile Include="JSValueTreeReader.cs" />
+    <Compile Include="JSValueTreeWriter.cs" />
+    <Compile Include="JSValueWriter.cs" />
+    <Compile Include="JSValueWriterGenerator.cs" />
+    <Compile Include="OneOf.cs" />
+    <Compile Include="ReactAttributes.cs" />
+    <Compile Include="ReactCallback.cs" />
+    <Compile Include="ReactConstantInfo.cs" />
+    <Compile Include="ReactConstantProvider.cs" />
+    <Compile Include="ReactConstantProviderInfo.cs" />
+    <Compile Include="ReactContext.cs" />
+    <Compile Include="ReactContextGenerator.cs" />
+    <Compile Include="ReactEnumerableExtensions.cs" />
+    <Compile Include="ReactEventInfo.cs" />
+    <Compile Include="ReactException.cs" />
+    <Compile Include="ReactFunctionInfo.cs" />
+    <Compile Include="ReactInitializerInfo.cs" />
+    <Compile Include="ReactMethodInfo.cs" />
+    <Compile Include="ReactModuleInfo.cs" />
+    <Compile Include="ReactPackageBuilderExtensions.cs" />
+    <Compile Include="ReactPackageProvider.cs" />
+    <Compile Include="ReactPromise.cs" />
+    <Compile Include="ReactSyncMethodInfo.cs" />
+    <Compile Include="ReflectionReactPackageProvider.cs" />
+    <Compile Include="ViewManagerAttributes.cs" />
+    <Compile Include="ViewManagerEvent.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="..\Microsoft.ReactNative.SharedManaged\*.cs" />
     <EmbeddedResource Include="Properties\Microsoft.ReactNative.Managed.rd.xml" />
   </ItemGroup>
   <ItemGroup>
