+++ vnext/Microsoft.ReactNative.Managed/OneOf.cs
