+++ vnext/Microsoft.ReactNative.Managed/ReactAttributes.cs
