+++ vnext/Microsoft.ReactNative.Managed/ReactCallback.cs
