+++ vnext/Microsoft.ReactNative.Managed/ReactConstantInfo.cs
