+++ vnext/Microsoft.ReactNative.Managed/ReactConstantProvider.cs
