+++ vnext/Microsoft.ReactNative.Managed/ReactConstantProviderInfo.cs
