+++ vnext/Microsoft.ReactNative.Managed/ReactContext.cs
