+++ vnext/Microsoft.ReactNative.Managed/ReactContextGenerator.cs
