+++ vnext/Microsoft.ReactNative.Managed/ReactEnumerableExtensions.cs
