+++ vnext/Microsoft.ReactNative.Managed/ReactEventInfo.cs
