+++ vnext/Microsoft.ReactNative.Managed/ReactException.cs
