+++ vnext/Microsoft.ReactNative.Managed/ReactFunctionInfo.cs
