+++ vnext/Microsoft.ReactNative.Managed/ReactInitializerInfo.cs
