+++ vnext/Microsoft.ReactNative.Managed/ReactMethodInfo.cs
