+++ vnext/Microsoft.ReactNative.Managed/ReactModuleInfo.cs
