+++ vnext/Microsoft.ReactNative.Managed/ReactPackageBuilderExtensions.cs
