+++ vnext/Microsoft.ReactNative.Managed/ReactPackageProvider.cs
