+++ vnext/Microsoft.ReactNative.Managed/ReactPromise.cs
