+++ vnext/Microsoft.ReactNative.Managed/ReactSyncMethodInfo.cs
