+++ vnext/Microsoft.ReactNative.Managed/ReflectionReactPackageProvider.cs
