+++ vnext/Microsoft.ReactNative.Managed/ViewManagerAttributes.cs
