+++ vnext/Microsoft.ReactNative.Managed/ViewManagerEvent.cs
