--- vnext/Microsoft.ReactNative.SharedManaged/Microsoft.ReactNative.SharedManaged.projitems
@@ -1,46 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>67a1076f-7790-4203-86ea-4402ccb5e782</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Microsoft.ReactNative.SharedManaged</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)AttributedViewManager.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IJSValueTreeReader.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactContext.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IReactPromise.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)JSValue.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)JSValueGenerator.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)JSValueReader.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)JSValueReaderGenerator.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)JSValueTreeReader.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)JSValueTreeWriter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)JSValueWriter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)JSValueWriterGenerator.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)OneOf.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactAttributes.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactCallback.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactConstantInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactConstantProvider.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactConstantProviderInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactContextGenerator.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactEnumerableExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactEventInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactException.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactFunctionInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactInitializerInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactMethodInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactModuleInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactPackageProvider.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactPackageBuilderExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactPromise.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReactSyncMethodInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ReflectionReactPackageProvider.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ViewManagerAttributes.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ViewManagerEvent.cs" />
-  </ItemGroup>
-</Project>
